# asset/swagger/openapi3.json
{"components":{"requestBodies":{"CreateUsersRequest":{"content":{"application/json":{"schema":{"properties":{"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"},"tag":{"type":"string"}},"type":"object"}}},"description":"Request used for creating a user.","required":true},"LoginUsersRequest":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"description":"Request used for logging a user.","required":true},"UpdateUsersRequest":{"content":{"application/json":{"schema":{"properties":{"active":{"format":"int32","type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"},"tag":{"type":"string"}},"type":"object"}}},"description":"Request used for updating a user.","required":true}},"responses":{"ClientErrorResponse":{"content":{"application/json":{"schema":{"properties":{"code":{"default":400,"format":"int32","type":"integer"},"error":{"type":"string"}}}}},"description":"Response when client errors happen."},"CreateUsersResponse":{"content":{"application/json":{"schema":{"properties":{"code":{"default":200,"format":"int32","type":"integer"},"user":{"$ref":"#/components/schemas/User"}}}}},"description":"Response returned back after creating users."},"DeleteUsersResponse":{"content":{"application/json":{"schema":{"properties":{"code":{"default":200,"format":"int32","type":"integer"},"user":{"type":"string"}}}}},"description":"Response returned back after deleting users."},"GetAllUsersResponse":{"content":{"application/json":{"schema":{"properties":{"code":{"default":200,"format":"int32","type":"integer"},"users":{"items":{"items":{"$ref":"#/components/schemas/User"}},"type":"array"}}}}},"description":"Response returned back after getting all users."},"LoginUsersResponse":{"content":{"application/json":{"schema":{"properties":{"code":{"default":200,"format":"int32","type":"integer"},"token":{"type":"string"}}}}},"description":"Response returned back after logging users."},"ServerErrorResponse":{"content":{"application/json":{"schema":{"properties":{"code":{"default":500,"format":"int32","type":"integer"},"error":{"type":"string"}}}}},"description":"Response when server errors happen."},"UpdateUsersResponse":{"content":{"application/json":{"schema":{"properties":{"code":{"default":200,"format":"int32","type":"integer"},"user":{"$ref":"#/components/schemas/User"}}}}},"description":"Response returned back after updating users."}},"schemas":{"Login":{"properties":{"email":{"default":"alonso12@gmail.com","type":"string"},"password":{"default":"123456","type":"string"}},"type":"object"},"User":{"properties":{"active":{"format":"int32","type":"integer"},"email":{"type":"string"},"id":{"format":"int64","type":"integer"},"name":{"type":"string"},"password":{"type":"string"},"tag":{"type":"string"}},"type":"object"}},"securitySchemes":{"bearerAuth":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}},"externalDocs":{"description":"Find out more about OpenAPI Specification.","url":"https://swagger.io/specification/"},"info":{"contact":{"email":"alonso.jorgeluis.01@gmail.com","name":"Jorge Luis Alonso","url":"https://github.com/Alonso9696"},"description":"REST APIs used for interacting with the ToDo Service.","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"title":"Global Hitss","version":"1.0.0"},"openapi":"3.0.0","paths":{"/signin":{"post":{"description":"Login user specifying email and password to get token.","operationId":"SignInUser","requestBody":{"$ref":"#/components/requestBodies/LoginUsersRequest"},"responses":{"200":{"$ref":"#/components/responses/LoginUsersResponse"},"400":{"$ref":"#/components/responses/ClientErrorResponse"},"500":{"$ref":"#/components/responses/ServerErrorResponse"}},"summary":"Sing in user.","tags":["Login"]}},"/users":{"get":{"description":"Gets all user rows.","operationId":"GetAllUser","responses":{"200":{"$ref":"#/components/responses/GetAllUsersResponse"},"400":{"$ref":"#/components/responses/ClientErrorResponse"},"500":{"$ref":"#/components/responses/ServerErrorResponse"}},"summary":"Get all users.","tags":["Users"]},"post":{"description":"Register a user with the requested fields.","operationId":"CreateUser","requestBody":{"$ref":"#/components/requestBodies/CreateUsersRequest"},"responses":{"200":{"$ref":"#/components/responses/CreateUsersResponse"},"400":{"$ref":"#/components/responses/ClientErrorResponse"},"500":{"$ref":"#/components/responses/ServerErrorResponse"}},"summary":"Create user.","tags":["Users"]}},"/users/{id}":{"delete":{"description":"Delete a user by id.","operationId":"DeleteUser","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int64","type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/DeleteUsersResponse"},"400":{"$ref":"#/components/responses/ClientErrorResponse"},"500":{"$ref":"#/components/responses/ServerErrorResponse"}},"summary":"Delete user.","tags":["Users"]},"put":{"description":"Updates a user with the requested fields.","operationId":"UpdateUser","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int64","type":"integer"}}],"requestBody":{"$ref":"#/components/requestBodies/UpdateUsersRequest"},"responses":{"200":{"$ref":"#/components/responses/UpdateUsersResponse"},"400":{"$ref":"#/components/responses/ClientErrorResponse"},"500":{"$ref":"#/components/responses/ServerErrorResponse"}},"summary":"Update user.","tags":["Users"]}}},"security":[{"bearerAuth":[]}],"servers":[{"description":"URL endpoints","url":"{Scheme}://{Host}:{Port}/api/v1","variables":{"Host":{"default":"localhost","enum":["localhost","192.168.1.5"]},"Port":{"default":"4000","enum":["4000","5000"]},"Scheme":{"default":"http","enum":["http","https"]}}}],"tags":[{"description":"Endpoints to manage users.","externalDocs":{"description":"external docs description","url":"https://swagger.io/specification/"},"name":"Users"},{"description":"Endpoint to login users.","externalDocs":{"description":"external docs description","url":"https://swagger.io/specification/"},"name":"Login"}]}